+++ cmd/pvegen/testgen.go
@@ -335,7 +335,7 @@ func renderBehaviorTestPreamble(builder *strings.Builder, pkgName string) {
 	builder.WriteString("\t\"strconv\"\n")
 	builder.WriteString("\t\"sync\"\n")
 	builder.WriteString("\t\"testing\"\n\n")
-	fmt.Fprintf(builder, "\t\"github.com/fivetwenty-io/proxmox-apiclient-go/v3/pkg/api/%s\"\n", pkgName)
+	fmt.Fprintf(builder, "\t\"github.com/fivetwenty-io/proxmox-apiclient-go/v3/%s/%s\"\n", activeDialect.pkgImportRoot, pkgName)
 	builder.WriteString("\tpveclient \"github.com/fivetwenty-io/proxmox-apiclient-go/v3/pkg/client\"\n")
 	builder.WriteString(")\n\n")